// api/src/cid/dto/resposta-cid.dto.ts
/** Resposta detalhe — BigInt como string no JSON. */
export class RespostaCidDto {
  id: string;
  codigo_cid: string | null;
  descricao: string | null;
  descricao_cap: string | null;
  descricao_categoria: string | null;
  observacao: string | null;
  id_usuario_auditoria: string | null;
  endereco_ip_auditoria: string | null;
  nome_aplicacao_auditoria: string | null;
}